Adding the rename changes I had in the new PR to make it clean and independent
