+++ scripts/tests/yaml/chiptool.py
