+++ scripts/tests/yaml/extensions/wildcard_response_extractor_cluster.py
