+++ scripts/tests/yaml/paths_finder.py
